Implemented try/catch to add possibilities
The response of the first service is honored.
If you don't want yours to be honored, throw an Ignored Exception and
the next in line will be picked up.

Exceptions can break loops or ignore them and continue the next one.
If a service validates to true or false, no other services will be
checked